Add a re-usable Action for syncing files
To make it easier to i.e. keep OpenAPI specifications in sync between
repos, we can create a dependency-less file syncing Action.

This uses the GitHub API - with a given GitHub token - to retrieve the
file, and then raise a PR if the file has changes.

This needs to take into account whether the PR is already open - and if
so `edit` it - otherwise `create` it.

We use `<code>` tags to avoid having to escape the backquotes.

We also add the step summary into the output, so we can see the URLs
more easily.